requirements.txt
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
tensorflow-hub==0.15.0
streamlit
numpy
matplotlib
pillow